<1mm acurravy, moving to goal to be tested

+++ include/helpers_broker.h
@@ -14,7 +14,7 @@ float get_orientation_float (long encoder1, long encoder2);
 float compute_linear_dist (const long encoder1, const long encoder2);
 float compute_linear_speed (long encoder1, long encoder2, long elapsed);
 int get_sensor_values (int fd, void * buf, ssize_t msg_size);
-void update_speed (const int immediate, int *current_speed, const int new_speed);
+void update_speed (const float immediate, float *current_speed, const float new_speed);
 float pseudo_gaussian_derivative(const int x);
 float obstacle_temporary_derivative (const int x);
 float apply_circular (float f (int), const int u, const int x);
